отделил компонент списка городов от главной страницы
